demo-server/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
channels
channels_redis
daphne
django-cors-headers
dj_database_url
djangorestframework
djangorestframework_simplejwt
django-phonenumber-field
django-environ
haversine
phonenumbers
Pillow
pyDes
pytz
redis
requests
sqlparse
urllib3
filelock
six
whitenoise
